[Draft, Add] Initial visualization abstraction
Each experiment might have different visualization needs. It should be
possible to abstract this in a way to allow for configurable
visualization.

The current idea consists of three main parts:
1. VisualizationBlock: Handles when visualization should happen (every n
batch/step)
2. Figure/Suplots: These hold the figure/subplots configuration
3. Plot: the object holding the plotting object/function and feeds data
   to it

The current main issues are how to give in args/kwargs to the plotting
function/objects and how to make eg. label/pred appear in the figure
titles.

For handling args/kwargs one might need to add two different target
arguments: target_args, target_kwargs. These can then be properly added
to the plotting call.

For adding labels to titles one might opt for a formatting strategi,
letting the title string be on the format: "label: {label}, pred:
{pred}" and use the string formatting to fill in.